working task6 ?

+++ user/user.h
@@ -25,7 +25,7 @@ int uptime(void);
 int memsize(void);
 void set_ps_priority(int);
 int  set_cfs_priority(int);
-void get_cfs_priority(int,uint64*,uint64*,uint64*,uint64*);
+void get_cfs_priority(int,int*,int*,int*,int*);
 
 // ulib.c
 int stat(const char*, struct stat*);
